Split project into multiple packages
