+++ core/js/src/main/scala/eu/joaocosta/minart/backend/defaults/package.scala
@@ -6,6 +6,6 @@ package object defaults {
   implicit val defaultCanvas: DefaultBackend[Any, HtmlCanvas] =
     DefaultBackend.fromFunction((_) => new HtmlCanvas())
 
-  implicit val defaultRenderLoop: DefaultBackend[Any, JsLoopRunner.type] =
+  implicit val defaultLoopRunner: DefaultBackend[Any, JsLoopRunner.type] =
     DefaultBackend.fromConstant(JsLoopRunner)
 }
