+++ core/jvm/src/main/scala/eu/joaocosta/minart/backend/AwtCanvas.scala
@@ -11,11 +11,12 @@ import javax.swing.JFrame
 import javax.swing.WindowConstants
 import scala.collection.JavaConverters._
 
-import eu.joaocosta.minart.core.Canvas.Resource
-import eu.joaocosta.minart.core.KeyboardInput.Key
 import eu.joaocosta.minart.core._
-
-import eu.joaocosta.minart.core.LowLevelCanvas.ExtendedSettings
+import eu.joaocosta.minart.graphics.Canvas.Resource
+import eu.joaocosta.minart.graphics.LowLevelCanvas.ExtendedSettings
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.input.KeyboardInput.Key
+import eu.joaocosta.minart.input._
 
 /** A low level Canvas implementation that shows the image in an AWT/Swing window.
   */
