+++ core/jvm/src/test/scala/eu/joaocosta/minart/backend/JavaRenderLoopSpec.scala
@@ -2,6 +2,7 @@ package eu.joaocosta.minart.backend
 
 import eu.joaocosta.minart.backend._
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.graphics._
 
 object JavaRenderLoopSpec extends RenderLoopTests {
   lazy val loopRunner: LoopRunner = JavaLoopRunner
