+++ core/native/src/main/scala/eu/joaocosta/minart/backend/SdlCanvas.scala
@@ -6,8 +6,10 @@ import scalanative.unsigned._
 import sdl2.Extras._
 import sdl2.SDL._
 
-import eu.joaocosta.minart.core.KeyboardInput.Key
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.input.KeyboardInput.Key
+import eu.joaocosta.minart.input._
 
 /** A low level Canvas implementation that shows the image in a SDL Window.
   */
