+++ core/native/src/main/scala/eu/joaocosta/minart/backend/SdlKeyMapping.scala
@@ -3,7 +3,7 @@ package eu.joaocosta.minart.backend
 import sdl2.Extras._
 import sdl2.SDL._
 
-import eu.joaocosta.minart.core.KeyboardInput._
+import eu.joaocosta.minart.input.KeyboardInput._
 
 object SdlKeyMapping extends KeyMapping[SDL_Keycode] {
   protected final val mappings: Map[Int, Key] = Map(
