+++ core/native/src/main/scala/eu/joaocosta/minart/backend/SdlLoopRunner.scala
@@ -10,6 +10,7 @@ import sdl2.SDL._
 
 import eu.joaocosta.minart.core._
 import eu.joaocosta.minart.core.Loop._
+import eu.joaocosta.minart.graphics.FrameRate
 
 object SdlLoopRunner extends LoopRunner {
   def finiteLoop[S](
