+++ core/shared/src/main/scala/eu/joaocosta/minart/backend/PpmCanvas.scala
@@ -1,6 +1,7 @@
 package eu.joaocosta.minart.backend
 
-import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.input._
 
 /** A low level Canvas implementation that outputs the image in the PPM format to the stdout.
   * This canvas doesn't support fetching the keyboard input.
