+++ core/shared/src/main/scala/eu/joaocosta/minart/graphics/Canvas.scala
@@ -1,6 +1,7 @@
-package eu.joaocosta.minart.core
+package eu.joaocosta.minart.graphics
 
 import eu.joaocosta.minart.backend.defaults.DefaultBackend
+import eu.joaocosta.minart.input._
 
 /** Canvas that can be painted.
   *
