+++ core/shared/src/main/scala/eu/joaocosta/minart/graphics/ImpureRenderLoop.scala
@@ -1,7 +1,8 @@
-package eu.joaocosta.minart.core
+package eu.joaocosta.minart.graphics
 
 import eu.joaocosta.minart.backend.defaults.DefaultBackend
-import eu.joaocosta.minart.core.RenderLoop._
+import eu.joaocosta.minart.core.LoopRunner
+import eu.joaocosta.minart.graphics.RenderLoop._
 
 object ImpureRenderLoop extends RenderLoop[Function1, Function2] {
   def finiteRenderLoop[S](
