+++ core/shared/src/main/scala/eu/joaocosta/minart/input/PointerInput.scala
@@ -1,6 +1,6 @@
-package eu.joaocosta.minart.core
+package eu.joaocosta.minart.input
 
-import eu.joaocosta.minart.core.KeyboardInput.Key
+import eu.joaocosta.minart.input.KeyboardInput.Key
 
 /** The pointer input stores the state of the mouse (or similar device) at a certain point in time.
   * It also accumulates points that have been pressed and released.
