+++ examples/fire/shared/src/main/scala/eu/joaocosta/minart/examples/Fire.scala
@@ -4,6 +4,8 @@ import scala.concurrent.duration._
 
 import eu.joaocosta.minart.backend.defaults._
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.input._
 
 object Fire {
   val canvasSettings = Canvas.Settings(width = 128, height = 128, scale = 4)
