+++ examples/settings/shared/src/main/scala/eu/joaocosta/minart/examples/Settings.scala
@@ -1,8 +1,9 @@
 package eu.joaocosta.minart.examples
 
 import eu.joaocosta.minart.backend.defaults._
-import eu.joaocosta.minart.core.KeyboardInput.Key
+import eu.joaocosta.minart.input.KeyboardInput.Key
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.graphics._
 
 object Settings {
   val settingsA = Canvas.Settings(width = 128, height = 128, scale = 4, clearColor = Color(128, 255, 128))
