+++ pure/shared/src/main/scala/eu/joaocosta/minart/core/pure/RIO.scala
@@ -1,4 +1,4 @@
-package eu.joaocosta.minart.pure
+package eu.joaocosta.minart.core.pure
 
 import scala.concurrent.Future
 import scala.util.Try
