+++ pure/shared/src/main/scala/eu/joaocosta/minart/graphics/pure/CanvasIO.scala
@@ -1,9 +1,12 @@
-package eu.joaocosta.minart.pure
+package eu.joaocosta.minart.graphics.pure
 
 import scala.concurrent.Future
 import scala.util.Try
 
 import eu.joaocosta.minart.core._
+import eu.joaocosta.minart.core.pure._
+import eu.joaocosta.minart.graphics._
+import eu.joaocosta.minart.input._
 
 /** Representation of a canvas operation, with the common Monad operations.
   */
