+++ pure/shared/src/test/scala/eu/joaocosta/minart/core/pure/RIOSpec.scala
@@ -1,4 +1,4 @@
-package eu.joaocosta.minart.pure
+package eu.joaocosta.minart.core.pure
 
 import scala.concurrent._
